+++ src/test/java/io/dataspaceconnector/model/app/AppFactoryTest.java
@@ -1,0 +1,149 @@
+/*
+ * Copyright 2020 Fraunhofer Institute for Software and Systems Engineering
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.dataspaceconnector.model.app;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.net.URI;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class AppFactoryTest {
+
+    private AppFactory factory;
+
+    @BeforeEach
+    public void init() {
+        this.factory = new AppFactory();
+    }
+
+    @Test
+    public void default_value_is_empty() {
+        /* ARRANGE */
+        // Nothing to arrange here.
+
+        /* ACT && ASSERT */
+        assertEquals("", AppFactory.DEFAULT_VALUE);
+    }
+
+    @Test
+    public void default_uri_is_app_com() {
+        /* ARRANGE */
+        // Nothing to arrange here.
+
+        /* ACT && ASSERT */
+        assertEquals(URI.create("https://app.com"), AppFactory.DEFAULT_URI);
+    }
+
+    @Test
+    public void create_nullDesc_throwIllegalArgumentException() {
+        /* ARRANGE */
+        // Nothing to arrange.
+
+        /* ACT && ASSERT */
+        assertThrows(IllegalArgumentException.class, () -> factory.create(null));
+    }
+
+    @Test
+    public void create_validDesc_creationDateNull() {
+        /* ARRANGE */
+        // Nothing to arrange.
+
+        /* ACT */
+        final var result = factory.create(new AppDesc());
+
+        /* ASSERT */
+        assertNull(result.getCreationDate());
+    }
+
+    @Test
+    public void create_validDesc_modificationDateNull() {
+        /* ARRANGE */
+        // Nothing to arrange.
+
+        /* ACT */
+        final var result = factory.create(new AppDesc());
+
+        /* ASSERT */
+        assertNull(result.getModificationDate());
+    }
+
+    @Test
+    public void create_validDesc_idNull() {
+        /* ARRANGE */
+        // Nothing to arrange.
+
+        /* ACT */
+        final var result = factory.create(new AppDesc());
+
+        /* ASSERT */
+        assertNull(result.getId());
+    }
+
+    @Test
+    public void update_sameDesc_returnFalse() {
+        /* ARRANGE */
+        final var app = factory.create(new AppDesc());
+
+        /* ACT */
+        final var result = factory.update(app, new AppDesc());
+
+        /* ASSERT */
+        Assertions.assertFalse(result);
+    }
+
+    @Test
+    public void update_differentLanguage_setLanguage() {
+        /* ARRANGE */
+        final var desc = new AppDesc();
+        desc.setLanguage("DE");
+
+        final var app = factory.create(new AppDesc());
+
+        /* ACT */
+        factory.update(app, desc);
+
+        /* ASSERT */
+        assertEquals(desc.getLanguage(), app.getLanguage());
+    }
+
+    @Test
+    public void update_sameLanguage_returnFalse() {
+        /* ARRANGE */
+        final var desc = new AppDesc();
+        desc.setLanguage("DE");
+
+        final var app = factory.create(desc);
+
+        /* ACT */
+        final var result = factory.update(app, desc);
+
+        /* ASSERT */
+        Assertions.assertFalse(result);
+    }
+
+    @Test
+    public void update_nullDesc_throwIllegalArgumentException() {
+        /* ARRANGE */
+        final var app = factory.create(new AppDesc());
+
+        /* ACT && ASSERT */
+        assertThrows(IllegalArgumentException.class, () -> factory.update(app, null));
+    }
+
+}